2020-3-7

+++ src/main/java/zhangdahu/community/controller/PublishController.java
@@ -26,7 +26,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    private String edit(@PathVariable(name="id") Integer id,
+    private String edit(@PathVariable(name="id") Long id,
                         Model model)
     {
         Question question=questionService.getById(id);
@@ -41,7 +41,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false)  Long id,
                             HttpServletRequest request,
                             Model model) {
         model.addAttribute("title",title);
